requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
SQLAlchemy==1.4.46
aiosqlite==0.17.0
python-dotenv==1.0.0